chcenged type in script age

+++ module_1_introduction_venv_conditional_instructions/homework/age.py
@@ -2,7 +2,7 @@ from datetime import datetime
 
 current_year = datetime.now().year
 
-birth_year = float(input("What is your year of birth? "))
+birth_year = int(input("What is your year of birth? "))
 
 years_old = int(current_year - birth_year)
